refactor(desktop): split persona_events tests and pending-retention helpers into modules
The B5 test additions pushed persona_events.rs (1202) and
commands/personas/mod.rs (1001) past the 1000-line file-size guard. Fix
the cause instead of adding exceptions: move the persona_events test
module to persona_events/tests.rs (the existing spawn_hash/types
pattern) and extract the retention-enqueue helpers
(retain_persona_pending / tombstone_persona_pending) into
commands/personas/pending.rs. Pure moves, no behavior change; visibility
widened only as far as the existing callers (commands::teams) require.
